requirements.txt
Django==4.2.4
python-decouple==3.5
dj_database_url==0.5.0
psycopg2==2.9.3
djangorestframework==3.14.0
pytest-django==4.5.2
